backend/requirements.txt
# Web Server & Framework
fastapi==0.135.3
uvicorn==0.44.0
starlette==1.0.0

# Database
sqlalchemy==2.0.49
psycopg2-binary==2.9.11
alembic==1.18.4
greenlet==3.3.2

# Settings & Environment
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2

# Security & Auth
bcrypt==3.2.0
PyJWT==2.3.0
cryptography==3.4.8

# Utilities
anyio==4.13.0
idna==3.11
python-dateutil==2.8.1
typing_extensions==4.15.0
click==8.3.2
Mako==1.3.10
MarkupSafe==3.0.3
h11==0.16.0
exceptiongroup==1.3.1
tomli==2.4.1
annotated-types==0.7.0
aiosmtplib==5.1.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
click==8.3.2
cryptography==46.0.7
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.3
fastapi-mail==1.6.2
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dotenv==1.2.2
regex==2025.11.3
SQLAlchemy==2.0.49
starlette==0.52.1
tomli==2.4.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
